tagger/requirements.txt
absl-py==0.1.12
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
enum34==1.1.6
Flask==0.12.2
gast==0.2.0
gevent==1.2.2
greenlet==0.4.13
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imutils==0.4.6
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.5
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.2
Pillow==5.0.0
protobuf==3.5.2.post1
PyYAML==3.12
redis==2.10.6
requests==2.18.4
scipy==1.0.0
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
